Tests suits for all with Order and Job1
